Update About.svelte
rewrite this code and content to sound more profesional and have a more impactful display,      <h2>Class located in Madison Wisconsin, with classes also offered online. Taught by Benjamin Kallman.</h2>
  <p class="title">Drawing and Painting. <br>Brush, pen, or digital. Guided classes either structured by students interest or standard class to find the students interest.</p>
  <p>Instructor: Benjamin Kallman<br>
Ben is an artist based in the midwest with a Bachelor’s degree in Fine Arts from The School of
the Art Institute of Chicago. He has been painting for 25 years.</p>
Introducing an exceptional opportunity to take Drawing and Painting classes taught by the talented Benjamin Kallman. Our classes are conveniently located in Madison, Wisconsin and also offered online for those who prefer a more flexible learning experience.

Our classes cater to a diverse range of interests, whether you prefer to work with brush, pen, or digital tools. We offer guided classes that are either structured by the student's interests or a standard class to help students discover their artistic passions.

Benjamin Kallman is an accomplished artist with a Bachelor's degree in Fine Arts from The School of the Art Institute of Chicago. He has been practicing painting for over 25 years and is excited to share his wealth of experience and knowledge with his students.

Join us today and unleash your artistic potential with the guidance of Benjamin Kallman.




Regenerate response